added tests for user routes
