--- Backend/src/app/__tests__/user/userSample.json
@@ -1,7 +1,0 @@
-{
-    "username": "fbellos",
-    "name": "Fidel",
-    "age": 23,
-    "email": "test@email.com",
-    "password": "123456"
-}
